Update plugin to use new handling system
Removed old version handling system and implemented new system.

+++ ColorCube-MC1_11/src/main/java/com/endercrest/colorcube/handler/MC1_11WorldBorderHandler.java
@@ -35,7 +35,7 @@ public class MC1_11WorldBorderHandler implements WorldBorderHandler {
         }
     }
 
-    public void setWorldBorder(Player player, Location origin, int radius) {
+    public void setWorldBorder(Player player, Location origin, double radius) {
         try {
             Constructor packetWorldBorderConstructor = NMSUtil.getNmsClass("PacketPlayOutWorldBorder").getConstructor(NMSUtil.getNmsClass("WorldBorder"), NMSUtil.getNmsClass("PacketPlayOutWorldBorder$EnumWorldBorderAction"));
